test: changed tests structure
