--- .github/workflows/python-tests.yml
@@ -1,46 +1,0 @@
-name: oodeel tests
-
-on:
-  push:
-    branches:
-      - master
-  pull_request:
-  schedule:
-    - cron: "0 2 * * 0" # Run tests every Sunday at 2am
-
-jobs:
-  checks:
-    runs-on: ubuntu-latest
-    strategy:
-      max-parallel: 4
-      matrix-tf:
-        include:
-          - python-version: 3.8
-            tf-version: 2.5
-          - python-version: 3.9
-            tf-version: 2.8
-          - python-version: "3.10"
-            tf-version: 2.11
-      matrix-torch:
-        include:
-          - python-version: 3.8
-            torch-version: 1.7
-          - python-version: 3.9
-            torch-version: 1.13
-          - python-version: "3.10"
-            torch-version: 2.0.0
-
-    steps:
-    - uses: actions/checkout@v3
-    - name: Set up Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Install dependencies
-      run: |
-        python -m pip install --upgrade pip
-        pip install tox
-    - name: Test with tox (Python ${{ matrix-tf.python-version }} - TensorFlow ${{ matrix-tf.tf-version }})
-      run: tox -e py$(echo ${{ matrix-tf.python-version }}-tf${{ matrix-tf.tf-version }} | tr -d .)
-    - name: Test with tox (Python ${{ matrix-torch.python-version }} - Pytorch ${{ matrix-torch.torch-version }})
-      run: tox -e py$(echo ${{ matrix-torch.python-version }}-tf${{ matrix-torch.torch-version }} | tr -d .)
