+++ tests/tensorflow/utils/test_tf_operator.py
